pyproject.toml
[tool.poetry]
name = "sod-compliance-agent"
version = "0.1.0"
description = "NetSuite SOD Compliance & Risk Assessment Multi-Agent Framework"
authors = ["Compliance Engineering Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# LangChain ecosystem
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langchain-anthropic = "^0.3.0"
langchain-community = "^0.3.0"
langgraph = "^0.2.0"

# Claude API
anthropic = "^0.39.0"

# Database
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"

# Embeddings (choose one or both)
voyageai = "^0.3.0"

# Task queue & caching
celery = "^5.3.4"
redis = "^5.0.1"

# API & Web
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"

# NetSuite integration
requests = "^2.31.0"
oauthlib = "^3.2.2"
requests-oauthlib = "^2.0.0"

# Notifications
sendgrid = "^6.11.0"
slack-sdk = "^3.27.0"

# Monitoring
prometheus-client = "^0.20.0"

# Utilities
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.6.0"
mypy = "^1.8.0"
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
omit = ["tests/*", "*/migrations/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
